add constant in separate file
